+++ src/BCC.MSBuildLog/BuildCrossCheckLogger.cs
@@ -29,18 +29,17 @@ namespace BCC.MSBuildLog
         public override void Initialize(IEventSource eventSource)
         {
             var environmentProvider = new EnvironmentProvider();
-            var fileSystem = new FileSystem();
 
             var baseUrl = environmentProvider.GetEnvironmentVariable("BCC_URL") ?? "https://buildcrosscheck.azurewebsites.net";
             var restClient = new RestClient(baseUrl);
-
-            var submissionService = new SubmissionService(fileSystem, restClient);
+            var submissionService = new SubmissionService(restClient);
 
             var buildServiceProvider = new BuildServiceProvider(environmentProvider);
             var buildService = buildServiceProvider.GetBuildService();
             var parameterParser = new ParameterParser(environmentProvider, buildService);
 
             var logDataBuilderFactory = new LogDataBuilderFactory();
+            var fileSystem = new FileSystem();
 
             Initialize(fileSystem, eventSource, environmentProvider, submissionService, parameterParser, logDataBuilderFactory);
         }
@@ -79,6 +78,12 @@ namespace BCC.MSBuildLog
             if (string.IsNullOrWhiteSpace(_parameters.Hash))
             {
                 _environmentProvider.WriteLine("BuildCrossCheck Hash is not present");
+                return;
+            }
+
+            if (!_parameters.PullRequestNumber.HasValue)
+            {
+                _environmentProvider.WriteLine("BuildCrossCheck PullRequestNumber is not present");
                 return;
             }
 
@@ -166,7 +171,7 @@ namespace BCC.MSBuildLog
 
                 var contents = createCheckRun.ToJson();
                 var bytes = Encoding.Unicode.GetBytes(contents);
-                submitSuccess = _submissionService.SubmitAsync(bytes, _parameters.Token, _parameters.Hash).Result;
+                submitSuccess = _submissionService.SubmitAsync(bytes, _parameters).Result;
             }
             catch (Exception exception)
             {