+++ src/BCC.MSBuildLog/Interfaces/IEnvironmentProvider.cs
@@ -4,6 +4,8 @@
     {
         string GetEnvironmentVariable(string name);
 
+        int? GetIntEnvironmentVariable(string name);
+
         void WriteLine(string line);
 
         void DebugLine(string line);